requirements.txt
git+https://github.com/titouanc/asyncirc
aiohttp==0.21
aioamqp
aioauth-client
humanize
python-dateutil
autobahn
beautifulsoup4
pytz
roman
paho-mqtt
